bcm283*: further refactoring
 * removed all global variables, replacing them with object properties and links
 * clean up interface between bcm2835_property and bcm2835_fb (all
   framebuffer reconfig logic is now local to bcm2835_fb)
 * add framebuffer properties for initial config (e.g. bcm2835_fb.pixo=0 for Windows)

+++ include/hw/arm/bcm2835.h
@@ -25,7 +25,7 @@ typedef struct BCM2835State {
     ARMCPU cpu;
     BCM2835PeripheralState peripherals;
 
-    uint64_t vcram_size;
+    uint32_t vcram_size;
 } BCM2835State;
 
 #endif /* BCM2835_H */
